feat(web): device detail edit/delete entry via shared DeviceEditModal (#57)
The device detail page had no edit/delete entry — changing a field meant
going back to the list, finding the row, using its action menu. The detail
page is the natural place to review and edit a device.

Extract the list page's inline device form into a shared <DeviceEditModal>
component (props: open, device, onSaved) so the list page and detail page
share one source of truth for the field set + validation. The list page
drops ~250 lines of form state/handlers/markup and now just toggles the
component open.

Detail page: the header gets admin-gated Edit/Delete buttons. Edit opens
the shared modal prefilled with the current device and refreshes the page
on save (onSaved=fetchDevice). Delete shows a ConfirmDialog (with the
device name), then DELETEs and navigates back to the list.

DeviceEditModal owns its own form state + Zod validation (deviceSchema) +
submit; a $effect hydrates the form whenever the modal opens (device=non-
null ⇒ edit, null ⇒ create). Callers just toggle open + supply the device.
